Re-enable strict compiler flags in unix port and adjust engine code base to conform

+++ src/math/vector2.c
@@ -5,7 +5,7 @@
 
 
 STATIC void vector2_class_print(const mp_print_t *print, mp_obj_t self_in, mp_print_kind_t kind){
-    ENGINE_PRINTF("%.15f %.15f", ((mp_obj_float_t)((vector2_class_obj_t*)self_in)->x).value, ((mp_obj_float_t)((vector2_class_obj_t*)self_in)->y).value);
+    ENGINE_PRINTF("%.09f %.09f", (double)((mp_obj_float_t)((vector2_class_obj_t*)self_in)->x).value, (double)((mp_obj_float_t)((vector2_class_obj_t*)self_in)->y).value);
 }
